@@ -19,21 +19,21 @@
     ]
   },
   "devDependencies": {
-    "@babel/core": "^7.20.2",
-    "@babel/preset-env": "^7.20.2",
+    "@babel/core": "^7.23.2",
+    "@babel/preset-env": "^7.23.2",
     "@symfony/webpack-encore": "^4.1.1",
-    "autoprefixer": "^10.2.6",
+    "autoprefixer": "^10.2.5",
     "css-loader": "^6.7.1",
     "favicons": "^7.0.2",
     "favicons-webpack-plugin": "^6.0.0-alpha.2",
-    "glob": "^8.0.3",
+    "glob": "^10.3.10",
     "mini-css-extract-plugin": "^2.6.1",
-    "postcss": "^8.3.5",
+    "postcss": "^8.4.31",
     "postcss-easing-gradients": "github:Ambient-Impact/postcss-easing-gradients",
     "postcss-loader": "^7.0.1",
     "sass": "^1.43.4",
     "sass-loader": "^13.1.0",
-    "webpack": "^5.74.0",
+    "webpack": "^5.88.2",
     "webpack-cli": "^4.10.0",
     "webpack-remove-empty-scripts": "^1.0.1"
   },
